works - this is a stepping satone to accepting all of the merge up to  7613

+++ jxs599/CABLE/dev/2020/CABLE-3.0forMPI_Xmas/science/radiation/cbl_spitter.F90
@@ -14,7 +14,7 @@ implicit none
 !re-decl input args
 integer :: mp
 real :: cpi
-real :: doy(mp)       ! day of year !typecast from integer
+integer :: doy(mp)       ! day of year !typecast from integer
 real :: coszen(mp)    ! cos(zenith angle of sun)
 real :: fsd(mp)       ! short wave down (positive) w/m^2
